Comp 2025! (#82)
This PR merges all changes from Spring 2025, comp, and a few earlier PRs, bringing our codebase fully up-to-date. There are a few lingering issues that are likely to get resolved by other PRs:

DepthAI refactoring --> we will hopefully move away from this
Sonar old code, testing, etc. --> sonar team will clean up their files
Async I/O for IVC logs --> this is a discussion to be had later and it doesn't really matter, and IVC should be reorganized in general, but this is an electrical thing
HSV old code --> should be resolved when we realize what we need for this spring, everything is currently there for reference

---------

Co-authored-by: Vedarsh Shah <vedarsh.shah@gmail.com>
Co-authored-by: Crush <crush@duke-robotics.com>
Co-authored-by: Oogway <oogway@duke-robotics.com>
Co-authored-by: viethungle-vt1401 <brandon.viethungRI2001@gmail.com>
Co-authored-by: Will Denton <willdenton@outlook.com>
Co-authored-by: Patrick Zheng <pzheng111@gmail.com>
Co-authored-by: sakofficialcode <sakofficialcode@users.noreply.github.com>
Co-authored-by: Patrick Zheng <PatrickZheng0@users.noreply.github.com>
Co-authored-by: Srinath Iyer <srinathiyer06@gmail.com>
Co-authored-by: Maxwell Lin <git@maxwellmlin.com>
Co-authored-by: Ahaan Shah <1ahaanshah@gmail.com>
Co-authored-by: Siddharth Kini <89208894+sakofficialcode@users.noreply.github.com>
Co-authored-by: Mathew Chu <mathew.chu@duke.edu>
Co-authored-by: Patrick Zheng <78308782+PatrickZheng0@users.noreply.github.com>